+++ src/modules/database/types/database-cluster-user.ts
@@ -1,5 +1,8 @@
+import { IDatabaseClusterMysqlUserSettings } from '.';
+
 export interface IDatabaseClusterUser {
   name: string;
   password: string;
   role: string;
+  mysql_settings?: IDatabaseClusterMysqlUserSettings;
 }